+++ src/spikeglx.py
@@ -144,8 +144,12 @@ class Reader:
         sglx_file = str(self.file_bin)
         if self.is_mtscomp:
             self._raw = mtscomp.Reader()
-            ch_file = self.ch_file or _get_companion_file(sglx_file, ".ch")
-            self._raw.open(self.file_bin, ch_file)
+            self.ch_file = (
+                _get_companion_file(sglx_file, ".ch")
+                if self.ch_file is None
+                else self.ch_file
+            )
+            self._raw.open(self.file_bin, self.ch_file)
             if self._raw.shape != (self.ns, self.nc):
                 ftsec = self._raw.shape[0] / self.fs
                 if not self.ignore_warnings:  # avoid the checks for streaming data
@@ -381,7 +385,7 @@ class Reader:
             self.file_bin = file_out
         return file_out
 
-    def decompress_file(self, keep_original=True, **kwargs):
+    def decompress_file(self, keep_original=True, file_ch=None, **kwargs):
         """
         Decompresses a mtscomp file
         :param keep_original: defaults True. If False, the original compressed file (input)
@@ -404,11 +408,14 @@ class Reader:
             self.file_bin = kwargs["out"]
         return kwargs["out"]
 
-    def decompress_to_scratch(self, scratch_dir=None):
+    def decompress_to_scratch(self, file_meta=None, file_ch=None, scratch_dir=None):
         """
         Decompresses the file to a temporary directory
         Copy over the metadata file
         """
+        if file_meta is None:
+            file_meta = Path(self.file_bin).with_suffix(".meta")
+        file_ch = file_ch if file_ch is not None else self.ch_file
         if scratch_dir is None:
             bin_file = Path(self.file_bin).with_suffix(".bin")
         else:
@@ -422,6 +429,7 @@ class Reader:
             _logger.info("File is compressed, decompressing to a temporary file...")
             self.decompress_file(
                 keep_original=True,
+                file_ch=file_ch,
                 out=bin_file.with_suffix(".bin_temp"),
                 check_after_decompress=False,
                 overwrite=True,