+++ src/tests/unit/test_waveforms.py
@@ -15,8 +15,17 @@ from neurowaveforms.model import generate_waveform
 from neuropixel import trace_header
 from ibldsp.fourier import fshift
 
-
 TEST_PATH = Path(__file__).parents[1].joinpath("fixtures")
+
+
+def _dummy_spike(ns):
+    ns = 100
+    w = np.zeros(ns)
+    w[0] = 1
+    w = np.fft.fftshift(w)
+    sos = scipy.signal.butter(3, 0.2, "low", output="sos")
+    w = scipy.signal.sosfiltfilt(sos, w)
+    return w
 
 
 def make_array_peak_through_tip():
@@ -248,14 +257,10 @@ class TestWaveformExtractorArray(unittest.TestCase):
         sig_lens = [100, 101]
         for sample_shift in sample_shifts:
             for sig_len in sig_lens:
-                spike = scipy.signal.morlet2(sig_len, 8.0, 2.0)
+                spike = _dummy_spike(sig_len)
                 spike = -np.fft.irfft(
                     np.fft.rfft(np.real(spike)) * np.exp(1j * 45 / 180 * np.pi)
                 )
-                import pdb
-
-                pdb.set_trace
-
                 spike2 = fshift(spike, sample_shift)
                 spike3, shift_computed = waveforms.wave_shift_corrmax(spike, spike2)
 
@@ -267,7 +272,7 @@ class TestWaveformExtractorArray(unittest.TestCase):
         fs = 30000
         # Resynch in time spike2 onto spike
         sample_shift_original = 0.323
-        spike = scipy.signal.morlet2(100, 8.5, 2.0)
+        spike = _dummy_spike(100)
         spike = -np.fft.irfft(
             np.fft.rfft(np.real(spike)) * np.exp(1j * 45 / 180 * np.pi)
         )
@@ -284,7 +289,7 @@ class TestWaveformExtractorArray(unittest.TestCase):
     def test_wave_shift_waveform(self):
         sample_shift_original = 15.32
         # Create peak channel spike
-        spike_peak = scipy.signal.morlet2(100, 8.5, 2.0)  # 100 time samples
+        spike_peak = _dummy_spike(100)  # 100 time samples
         spike_peak = -np.fft.irfft(
             np.fft.rfft(np.real(spike_peak)) * np.exp(1j * 45 / 180 * np.pi)
         )
@@ -441,3 +446,29 @@ def test_wiggle():
     waveforms.plot_wiggle(wav, scale=40 * 1e-6, ax=ax[0])
     waveforms.double_wiggle(wav, scale=40 * 1e-6, fs=30_000, ax=ax[1])
     plt.close("all")
+
+
+def test_waveform_indices():
+    wxy, inds = waveforms.get_waveforms_coordinates(
+        trace_indices=np.array([22, 150, 370]), return_indices=True
+    )
+    np.testing.assert_array_equal(
+        wxy.flatten()[::19],
+        np.array(
+            [
+                59.0,
+                140.0,
+                59.0,
+                320.0,
+                11.0,
+                1400.0,
+                43.0,
+                1580.0,
+                27.0,
+                3580.0,
+                59.0,
+                3760.0,
+                np.nan,
+            ]
+        ),
+    )